First working prototype to visualize layer activations

+++ glassimaging/dataloading/transforms/totensor.py
@@ -7,8 +7,8 @@ class ToTensor(object):
 
     def __call__(self, sample):
         image = sample['data']
-        sample['data'] = torch.from_numpy(image)
+        sample['data'] = torch.from_numpy(image).float()
         if 'seg' in sample:
             segmentation = sample['seg']
-            sample['seg'] = torch.from_numpy(segmentation)
+            sample['seg'] = torch.from_numpy(segmentation).long()
         return sample
